betenda_api/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
Markdown==3.4.3
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.18.0
user-agents==2.2.0
zope.interface==6.0